memory_system/requirements.txt
# 核心依赖
chromadb>=0.4.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
click>=8.1.0
pydantic>=2.0.0

# 文档处理
pypdf>=3.0.0
python-docx>=1.0.0
markdown>=3.4.0

# LLM集成
openai>=1.50.0
python-dotenv>=1.0.0

# 工具
watchdog>=3.0.0  # 文件监听
httpx>=0.24.0    # HTTP客户端

# 可选依赖（高级功能）
# neo4j>=5.0.0          # 图数据库（阶段2）
# spacy>=3.5.0          # NER（阶段2）
# transformers>=4.30.0  # 深度学习（阶段2）
